+++ include/maze_path.hpp
@@ -5,12 +5,12 @@ namespace MazeGen
 {
 struct PathNode
 {
-    size_t cell;
+    int cell;
     Direction fromDirection;
     Direction toDirection;
 };
 
 typedef std::vector<PathNode> Path;
 
-Path FindPath(Maze &m, size_t from, size_t to);
+Path FindPath(Maze &m, int from, int to);
 }
